+++ lang/uk/passwords.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Password Reset Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are the default lines which match reasons
+    | that are given by the password broker for a password update attempt
+    | outcome such as failure due to an invalid password / reset token.
+    |
+    */
+
+    'reset'     => 'Ваш пароль змінено.',
+    'sent'      => 'Ми надіслали посилання для відновлення паролю на вашу електронну адресу.',
+    'throttled' => 'Зачекайте, перш ніж повторювати спробу.',
+    'token'     => 'Це посилання для відновлення паролю недійсне.',
+    'user'      => 'Користувача з такою електронною адресою не знайдено.',
+
+];