+++ lang/uk/validation.php
@@ -1,0 +1,194 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Validation Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines contain the default error messages used by
+    | the validator class. Some of these rules have multiple versions such
+    | as the size rules. Feel free to tweak each of these messages here.
+    |
+    */
+
+    'accepted' => 'Поле :attribute має бути прийняте.',
+    'accepted_if' => 'Поле :attribute має бути прийняте, коли :other є :value.',
+    'active_url' => 'Поле :attribute має містити дійсну URL-адресу.',
+    'after' => 'Поле :attribute має бути датою після :date.',
+    'after_or_equal' => 'Поле :attribute має бути датою після або рівною :date.',
+    'alpha' => 'Поле :attribute може містити лише літери.',
+    'alpha_dash' => 'Поле :attribute може містити лише літери, цифри, тире та підкреслення.',
+    'alpha_num' => 'Поле :attribute може містити лише літери та цифри.',
+    'array' => 'Поле :attribute має бути масивом.',
+    'ascii' => 'Поле :attribute може містити лише однобайтові буквено-цифрові символи та знаки.',
+    'before' => 'Поле :attribute має бути датою до :date.',
+    'before_or_equal' => 'Поле :attribute має бути датою до або рівною :date.',
+    'between' => [
+        'array' => 'Поле :attribute має містити від :min до :max елементів.',
+        'file' => 'Поле :attribute має бути від :min до :max кілобайт.',
+        'numeric' => 'Поле :attribute має бути від :min до :max.',
+        'string' => 'Поле :attribute має бути від :min до :max символів.',
+    ],
+    'boolean' => 'Поле :attribute має бути true або false.',
+    'can' => 'Поле :attribute містить неавторизоване значення.',
+    'confirmed' => 'Поле :attribute не співпадає з підтвердженням.',
+    'contains' => 'Поле :attribute не містить обов\'язкового значення.',
+    'current_password' => 'Невірний пароль.',
+    'date' => 'Поле :attribute має бути дійсною датою.',
+    'date_equals' => 'Поле :attribute має бути датою, що дорівнює :date.',
+    'date_format' => 'Поле :attribute має відповідати формату :format.',
+    'decimal' => 'Поле :attribute має мати :decimal знаків після коми.',
+    'declined' => 'Поле :attribute має бути відхилене.',
+    'declined_if' => 'Поле :attribute має бути відхилене, коли :other є :value.',
+    'different' => 'Поля :attribute та :other мають відрізнятися.',
+    'digits' => 'Поле :attribute має містити :digits цифр.',
+    'digits_between' => 'Поле :attribute має містити від :min до :max цифр.',
+    'dimensions' => 'Поле :attribute має неприпустимі розміри зображення.',
+    'distinct' => 'Поле :attribute має повторюване значення.',
+    'doesnt_end_with' => 'Поле :attribute не може закінчуватися одним з таких значень: :values.',
+    'doesnt_start_with' => 'Поле :attribute не може починатися з одного з таких значень: :values.',
+    'email' => 'Поле :attribute має бути дійсною email-адресою.',
+    'ends_with' => 'Поле :attribute має закінчуватися одним з таких значень: :values.',
+    'enum' => 'Обране значення для :attribute є неприпустимим.',
+    'exists' => 'Обране значення для :attribute є неприпустимим.',
+    'extensions' => 'Поле :attribute має мати одне з таких розширень: :values.',
+    'file' => 'Поле :attribute має бути файлом.',
+    'filled' => 'Поле :attribute має містити значення.',
+    'gt' => [
+        'array' => 'Поле :attribute має містити більше ніж :value елементів.',
+        'file' => 'Поле :attribute має бути більше ніж :value кілобайт.',
+        'numeric' => 'Поле :attribute має бути більше ніж :value.',
+        'string' => 'Поле :attribute має бути більше ніж :value символів.',
+    ],
+    'gte' => [
+        'array' => 'Поле :attribute має містити :value елементів або більше.',
+        'file' => 'Поле :attribute має бути більше або дорівнювати :value кілобайт.',
+        'numeric' => 'Поле :attribute має бути більше або дорівнювати :value.',
+        'string' => 'Поле :attribute має бути більше або дорівнювати :value символів.',
+    ],
+    'hex_color' => 'Поле :attribute має бути дійсним шістнадцятковим кодом кольору.',
+    'image' => 'Поле :attribute має бути зображенням.',
+    'in' => 'Обране значення для :attribute є неприпустимим.',
+    'in_array' => 'Поле :attribute має існувати в :other.',
+    'integer' => 'Поле :attribute має бути цілим числом.',
+    'ip' => 'Поле :attribute має бути дійсною IP-адресою.',
+    'ipv4' => 'Поле :attribute має бути дійсною IPv4-адресою.',
+    'ipv6' => 'Поле :attribute має бути дійсною IPv6-адресою.',
+    'json' => 'Поле :attribute має бути дійсним JSON-рядком.',
+    'list' => 'Поле :attribute має бути списком.',
+    'lowercase' => 'Поле :attribute має бути в нижньому регістрі.',
+    'lt' => [
+        'array' => 'Поле :attribute має містити менше ніж :value елементів.',
+        'file' => 'Поле :attribute має бути менше ніж :value кілобайт.',
+        'numeric' => 'Поле :attribute має бути менше ніж :value.',
+        'string' => 'Поле :attribute має бути менше ніж :value символів.',
+    ],
+    'lte' => [
+        'array' => 'Поле :attribute не повинно містити більше ніж :value елементів.',
+        'file' => 'Поле :attribute має бути менше або дорівнювати :value кілобайт.',
+        'numeric' => 'Поле :attribute має бути менше або дорівнювати :value.',
+        'string' => 'Поле :attribute має бути менше або дорівнювати :value символів.',
+    ],
+    'mac_address' => 'Поле :attribute має бути дійсною MAC-адресою.',
+    'max' => [
+        'array' => 'Поле :attribute не повинно містити більше ніж :max елементів.',
+        'file' => 'Поле :attribute не повинно бути більше ніж :max кілобайт.',
+        'numeric' => 'Поле :attribute не повинно бути більше ніж :max.',
+        'string' => 'Поле :attribute не повинно бути більше ніж :max символів.',
+    ],
+    'max_digits' => 'Поле :attribute не повинно містити більше ніж :max цифр.',
+    'mimes' => 'Поле :attribute має бути файлом типу: :values.',
+    'mimetypes' => 'Поле :attribute має бути файлом типу: :values.',
+    'min' => [
+        'array' => 'Поле :attribute має містити щонайменше :min елементів.',
+        'file' => 'Поле :attribute має бути щонайменше :min кілобайт.',
+        'numeric' => 'Поле :attribute має бути щонайменше :min.',
+        'string' => 'Поле :attribute має бути щонайменше :min символів.',
+    ],
+    'min_digits' => 'Поле :attribute має містити щонайменше :min цифр.',
+    'missing' => 'Поле :attribute має бути відсутнім.',
+    'missing_if' => 'Поле :attribute має бути відсутнім, коли :other є :value.',
+    'missing_unless' => 'Поле :attribute має бути відсутнім, якщо :other не є :value.',
+    'missing_with' => 'Поле :attribute має бути відсутнім, коли присутнє :values.',
+    'missing_with_all' => 'Поле :attribute має бути відсутнім, коли присутні :values.',
+    'multiple_of' => 'Поле :attribute має бути кратним :value.',
+    'not_in' => 'Обране значення для :attribute є неприпустимим.',
+    'not_regex' => 'Формат поля :attribute є неприпустимим.',
+    'numeric' => 'Поле :attribute має бути числом.',
+    'password' => [
+        'letters' => 'Поле :attribute має містити принаймні одну літеру.',
+        'mixed' => 'Поле :attribute має містити принаймні одну велику та одну малу літери.',
+        'numbers' => 'Поле :attribute має містити принаймні одну цифру.',
+        'symbols' => 'Поле :attribute має містити принаймні один символ.',
+        'uncompromised' => 'Вказане :attribute з\'явилося у витоку даних. Будь ласка, оберіть інше :attribute.',
+    ],
+    'present' => 'Поле :attribute має бути присутнім.',
+    'present_if' => 'Поле :attribute має бути присутнім, коли :other є :value.',
+    'present_unless' => 'Поле :attribute має бути присутнім, якщо :other не є :value.',
+    'present_with' => 'Поле :attribute має бути присутнім, коли присутнє :values.',
+    'present_with_all' => 'Поле :attribute має бути присутнім, коли присутні :values.',
+    'prohibited' => 'Поле :attribute заборонене.',
+    'prohibited_if' => 'Поле :attribute заборонене, коли :other є :value.',
+    'prohibited_unless' => 'Поле :attribute заборонене, якщо :other не знаходиться в :values.',
+    'prohibits' => 'Поле :attribute забороняє присутність :other.',
+    'regex' => 'Формат поля :attribute є неприпустимим.',
+    'required' => 'Поле :attribute обов\'язкове для заповнення.',
+    'required_array_keys' => 'Поле :attribute має містити записи для: :values.',
+    'required_if' => 'Поле :attribute обов\'язкове для заповнення, коли :other є :value.',
+    'required_if_accepted' => 'Поле :attribute обов\'язкове для заповнення, коли :other прийнято.',
+    'required_if_declined' => 'Поле :attribute обов\'язкове для заповнення, коли :other відхилено.',
+    'required_unless' => 'Поле :attribute обов\'язкове для заповнення, якщо :other не знаходиться в :values.',
+    'required_with' => 'Поле :attribute обов\'язкове для заповнення, коли присутнє :values.',
+    'required_with_all' => 'Поле :attribute обов\'язкове для заповнення, коли присутні :values.',
+    'required_without' => 'Поле :attribute обов\'язкове для заповнення, коли відсутнє :values.',
+    'required_without_all' => 'Поле :attribute обов\'язкове для заповнення, коли відсутні всі :values.',
+    'same' => 'Поле :attribute має співпадати з :other.',
+    'size' => [
+        'array' => 'Поле :attribute має містити :size елементів.',
+        'file' => 'Поле :attribute має бути :size кілобайт.',
+        'numeric' => 'Поле :attribute має бути :size.',
+        'string' => 'Поле :attribute має бути :size символів.',
+    ],
+    'starts_with' => 'Поле :attribute має починатися з одного з таких значень: :values.',
+    'string' => 'Поле :attribute має бути рядком.',
+    'timezone' => 'Поле :attribute має бути дійсною часовою зоною.',
+    'unique' => 'Таке значення поля :attribute вже існує.',
+    'uploaded' => 'Не вдалося завантажити :attribute.',
+    'uppercase' => 'Поле :attribute має бути в верхньому регістрі.',
+    'url' => 'Поле :attribute має бути дійсною URL-адресою.',
+    'ulid' => 'Поле :attribute має бути дійсним ULID.',
+    'uuid' => 'Поле :attribute має бути дійсним UUID.',
+
+    /*
+    |--------------------------------------------------------------------------
+    | Custom Validation Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | Here you may specify custom validation messages for attributes using the
+    | convention "attribute.rule" to name the lines. This makes it quick to
+    | specify a specific custom language line for a given attribute rule.
+    |
+    */
+
+    'custom' => [
+        'attribute-name' => [
+            'rule-name' => 'власне-повідомлення',
+        ],
+    ],
+
+    /*
+    |--------------------------------------------------------------------------
+    | Custom Validation Attributes
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are used to swap our attribute placeholder
+    | with something more reader friendly such as "E-Mail Address" instead
+    | of "email". This simply helps us make our message more expressive.
+    |
+    */
+
+    'attributes' => [],
+
+];